+++ tests/integration/test_charm.py
@@ -37,6 +37,7 @@ async def test_image_relation(app: Application, test_charm: Application):
 
 
 @pytest.mark.asyncio
+@pytest.mark.abort_on_fail
 async def test_cos_agent_relation(app: Application):
     """
     arrange: An active charm.
@@ -55,6 +56,7 @@ async def test_cos_agent_relation(app: Application):
 
 
 @pytest.mark.asyncio
+@pytest.mark.abort_on_fail
 async def test_build_image(
     openstack_connection: Connection,
     dispatch_time: datetime,
@@ -70,6 +72,7 @@ async def test_build_image(
 
 # Ignore the "too many arguments" warning, as this is not significant for a test function where
 # the arguments are fixtures and the function is not expected to be called directly.
+@pytest.mark.abort_on_fail
 async def test_charm_another_app_does_not_rebuild_image(  # pylint: disable=R0913,R0917
     app: Application,
     test_charm: Application,
@@ -128,6 +131,7 @@ async def test_charm_another_app_does_not_rebuild_image(  # pylint: disable=R091
 
 
 @pytest.mark.asyncio
+@pytest.mark.abort_on_fail
 async def test_periodic_rebuilt(
     app: Application,
     app_config: dict,
@@ -179,6 +183,7 @@ async def _change_cronjob_to_minutes(unit: Unit, current_hour_interval: int):
 
 
 @pytest.mark.asyncio
+@pytest.mark.abort_on_fail
 async def test_log_rotated(app: Application):
     """
     arrange: A deployed active charm and manually write something to the log file.